# concierge/api/openapi.yaml
openapi: 3.0.3
info:
  title: concierge — Basis SSO admission & control plane API
  version: "1.0.0"
  description: >
    Wire-compatible Go rewrite of the Basis SSO admission broker
    (the former Basis SSO broker). Endpoint paths, JSON field names, and
    status codes are load-bearing: they are consumed by unmodified C# Unity
    clients and the unmodified C# UDP game server. See
    docs/concierge/research-sso-broker.md and docs/concierge/design.md for
    the source specification. HTML/plain-text/raw-file responses (enroll and
    join pages, /admin/client-config/{serverId}) are documented here for
    completeness but are written directly by the handlers rather than
    generated, since oapi-codegen's std-http-server mode only generates
    routing and parameter binding, not response encoding. Managed WebGL
    GameServers may terminate TLS using the read-only Kubernetes Secret
    configured by Broker.Kubernetes; that deployment-only setting is not
    mutable through this API.
paths:
...
  /web-client-config/{serverId}:
    get:
      operationId: getWebClientConfig
      parameters:
        - $ref: "#/components/parameters/ServerId"
      responses:
        "200":
          description: Browser client configuration with broker token relay endpoints.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientConfiguration"
        "404":
          description: Unknown server.
        "503":
          description: Web SSO is not configured.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
  /web-oidc/{serverId}/{providerId}/token:
    options:
      operationId: webOidcOptions
      parameters:
        - $ref: "#/components/parameters/ServerId"
        - $ref: "#/components/parameters/ProviderId"
      responses:
        "204":
          description: CORS preflight accepted.
        "403":
          description: Origin is not allowed.
    post:
      operationId: webOidcToken
      parameters:
        - $ref: "#/components/parameters/ServerId"
        - $ref: "#/components/parameters/ProviderId"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties:
                type: string
      responses:
        "200":
          description: Upstream OIDC token response.
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Invalid or unsupported grant request.
        "403":
          description: Origin or redirect URI is not allowed.
        "404":
          description: Unknown server/provider.
        "503":
          description: Web provider credentials are incomplete.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
  /join/{token}/web-config:
    get:
      operationId: joinWebConfig
      parameters:
        - $ref: "#/components/parameters/Token"
      responses:
        "200":
          description: Browser client configuration for the invited meeting.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientConfiguration"
        "404":
          description: Unknown invite token.
        "403":
          description: Origin is not allowed.
        "503":
          description: Web SSO configuration incomplete.
  /join/{token}/web-manifest:
    get:
      operationId: joinWebManifest
      parameters:
        - $ref: "#/components/parameters/Token"
      responses:
        "200":
          description: WebGL meeting manifest.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebMeetingManifest"
        "403":
          description: Origin is not allowed.
        "404":
          description: Unknown invite token.
        "409":
          description: Meeting is not ready.
  /join/{token}/details:
    get:
      operationId: joinDetails
      parameters:
        - $ref: "#/components/parameters/Token"
      responses:
        "200":
          description: Participant page details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingDetails"
        "404":
          description: Unknown invite token.
        "409":
          description: Meeting is not ready.
...
  /admin/servers:
    get:
      operationId: listAdminServers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminServerInfo"
        "401":
          description: Unauthorized.
  /admin/organization:
    get:
      operationId: getOrganization
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationOptions"
        "401":
          description: Unauthorized.
    put:
      operationId: putOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationOptions"
      responses:
        "204":
          description: Saved.
        "400":
          description: Structurally invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized.
...
  /admin/servers/{serverId}:
    put:
      operationId: putAdminServer
      parameters:
        - $ref: "#/components/parameters/ServerId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminServerWrite"
      responses:
        "204":
          description: Upserted.
        "400":
          description: Structurally invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized.
    delete:
      operationId: deleteAdminServer
      parameters:
        - $ref: "#/components/parameters/ServerId"
      responses:
        "204":
          description: Deleted.
        "401":
          description: Unauthorized.
        "404":
          description: Unknown server.
...
components:
  parameters:
    ServerId:
      name: serverId
      in: path
      required: true
      schema:
        type: string
    Token:
      name: token
      in: path
      required: true
      schema:
        type: string
    ProviderId:
      name: providerId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ProblemDetails:
      type: object
      description: RFC 7807 application/problem+json envelope, matching ASP.NET Core's Results.Problem default shape.
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
...
    ProviderOptions:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        issuer:
          type: string
        audience:
          type: string
        clientSecret:
          type: string
        webClientId:
          type: string
        webClientSecret:
          type: string
        tokenEndpoint:
          type: string
        jwksUri:
          type: string
        allowedHostedDomains:
          type: array
          items:
            type: string
        allowedGroups:
          type: array
          items:
            type: string
    OrganizationOptions:
      type: object
      properties:
        displayName:
          type: string
        defaultProviderId:
          type: string
        providers:
          type: array
          items:
            $ref: "#/components/schemas/ProviderOptions"
    AdminServerInfo:
      type: object
      required: [id, ready, providers, hasTicketSigningKey, hasTransportPublicKey]
      properties:
        id:
          type: string
        ticketSigningKeyEnvironmentVariable:
          type: string
        transportPublicKeyEnvironmentVariable:
          type: string
        providers:
          type: array
          items:
            $ref: "#/components/schemas/ProviderOptions"
        ready:
          type: boolean
        hasTicketSigningKey:
          type: boolean
        hasTransportPublicKey:
          type: boolean
        webSocketUri:
          type: string
        serverInfoUri:
          type: string
    AdminServerWrite:
      type: object
      description: Body for PUT /admin/servers/{serverId} (BrokerServerOptions equivalent).
      properties:
        id:
          type: string
        ticketSigningKeyEnvironmentVariable:
          type: string
        transportPublicKeyEnvironmentVariable:
          type: string
        ticketSigningKey:
          type: string
        transportPublicKey:
          type: string
        webSocketUri:
          type: string
        serverInfoUri:
          type: string
        providers:
          type: array
          items:
            $ref: "#/components/schemas/ProviderOptions"
...
    ServerTransportConfig:
      type: object
      properties:
        serverPublicKey:
          type: string
        admissionEndpoint:
          type: string
        allowUntrustedLoopbackCertificate:
          type: boolean
        webSocketUri:
          type: string
        serverInfoUri:
          type: string
    ClaimRule:
      type: object
      properties:
        claim:
          type: string
        values:
          type: array
          items:
            type: string
    AccessConfig:
      type: object
      properties:
        allowedGroups:
          type: array
          items:
            type: string
        allowedClaims:
          type: array
          items:
            $ref: "#/components/schemas/ClaimRule"
    ClientProviderConfig:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        issuer:
          type: string
        clientId:
          type: string
        clientSecret:
          type: string
        tokenEndpoint:
          type: string
        extraAuthParams:
          type: object
          additionalProperties:
            type: string
        scopes:
          type: array
          items:
            type: string
        displayNameClaims:
          type: array
          items:
            type: string
        access:
          $ref: "#/components/schemas/AccessConfig"
    RedirectConfig:
      type: object
      properties:
        mode:
          type: string
        host:
          type: string
        port:
          type: integer
        path:
          type: string
    EnforcementConfig:
      type: object
      properties:
        allowOfflineWithinTokenValidity:
          type: boolean
    MeetingDetails:
      type: object
      required: [title, webJoinUrl, nativeBridgeUrl]
      properties:
        title:
          type: string
        webJoinUrl:
          type: string
        nativeBridgeUrl:
          type: string
    WebMeetingManifest:
      type: object
      required: [configUrl, websocketUri, userName, password]
      properties:
        configUrl:
          type: string
        websocketUri:
          type: string
        userName:
          type: string
        password:
          type: string
    ClientConfiguration:
      type: object
      properties:
        defaultProviderId:
          type: string
        serverTransport:
          $ref: "#/components/schemas/ServerTransportConfig"
        providers:
          type: array
          items:
            $ref: "#/components/schemas/ClientProviderConfig"
        redirect:
          $ref: "#/components/schemas/RedirectConfig"
        enforcement:
          $ref: "#/components/schemas/EnforcementConfig"
...